requirements.txt
aiogram>=3.10
python-dotenv>=1.0
aiohttp>=3.9.0